+++ tests/views/customer_dashboards/test_customer_dashboard_leads_view.py
@@ -5,9 +5,9 @@ import pytest
 
 
 @pytest.mark.usefixtures('db_transaction', 'create_dependencies', 'login')
-class TestCustomerDashboardView(BaseDashboardTestView):
-    """Test customer dashboards."""
+class TestCustomerDashboardLeadsView(BaseDashboardTestView):
+    """Test customer dashboard for Leads."""
 
     base_paths = (
-        '/dashboards/customer/order',
+        '/dashboards/customer/lead',
     )