+++ tests/views/customer_dashboards/test_customer_dashboard_orders_view.py
@@ -1,0 +1,13 @@
+"""Test Customer Dashboards view."""
+from conftest import BaseDashboardTestView
+
+import pytest
+
+
+@pytest.mark.usefixtures('db_transaction', 'create_dependencies', 'login')
+class TestCustomerDashboardOrderView(BaseDashboardTestView):
+    """Test customer dashboard for Orders."""
+
+    base_paths = (
+        '/dashboards/customer/order',
+    )